restructor main componengt
