+++ package.json
@@ -26,7 +26,7 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.1",
+    "@angular-devkit/build-angular": "^13.3.11",
     "@angular/cli": "~13.2.1",
     "@angular/compiler-cli": "~13.2.0",
     "@types/jasmine": "~3.10.0",
