+++ projects/ag-rform/src/lib/ag-rform.module.ts
@@ -3,7 +3,6 @@ import { AgRFormComponent } from './ag-rform.component';
 import { FormFieldDirective } from './directives/form-field.directive';
 import { InputTagComponent } from './base/input-tag/input-tag.component';
 import { SelectTagComponent } from './base/select-tag/select-tag.component';
-import { AgrMainFormComponent } from './base/agr-main-form/agr-main-form.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
@@ -14,19 +13,17 @@ import { CommonModule } from '@angular/common';
     FormFieldDirective,
     InputTagComponent,
     SelectTagComponent,
-    AgrMainFormComponent,
   ],
   imports: [BrowserModule, ReactiveFormsModule, FormsModule, CommonModule],
   exports: [
     AgRFormComponent,
     InputTagComponent,
-    AgrMainFormComponent,
     FormFieldDirective,
   ],
   entryComponents: [
+    AgRFormComponent,
     InputTagComponent,
-    AgrMainFormComponent,
     FormFieldDirective,
   ],
 })
-export class AgRFormModule {}
+export class AgRFormModule { }
