+++ projects/ag-rform/src/lib/agr-form.component.html
@@ -21,4 +21,3 @@
       </ng-container>
     </div>
   </form>
-
