+++ projects/ag-rform/src/public-api.ts
@@ -5,8 +5,10 @@
 export * from './lib/ag-rform.service';
 export * from './lib/ag-rform.component';
 export * from './lib/ag-rform.module';
-export * from './lib/base/agrFormFieldComponent';
-export * from './lib/base/agr-main-form/agr-main-form.component';
-export * from './lib/base/input-tag/input-tag.component';
 export * from './lib/directives/form-field.directive';
 export * from './lib/mappers/formfield.mapper';
+export * from './lib/agrFormFieldComponent';
+//component list
+
+export * from './lib/ag-rform.component';
+export * from './lib/base/input-tag/input-tag.component';
