+++ src/app/app.component.html
@@ -1,8 +1,8 @@
-<lib-agr-main-form
+<lib-agr-form
   #reusableForm
   [fields]="field"
   [rowHeight]="'50px'"
   [cols]="12"
   (submit)="submit($event)"
   [data]=""
-></lib-agr-main-form>
+></lib-agr-form>
